requirements.txt
certifi==2019.6.16
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-heroku==0.3.1
django-mpesapy==0.1.2
django-registration==2.4.1
djangorestframework==3.9.4
idna==2.8
Pillow==6.0.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
requests==2.22.0
urllib3==1.25.3
whitenoise==4.1.2
